backend/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-jazzmin>=2.6.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
